refactor(cli): move code.
