+++ README.md
@@ -75,29 +75,30 @@ COMMANDS:
 
 GLOBAL OPTIONS:
    --domains value, -d value   Add a domain to the process. Can be specified multiple times.
-   --csr value, -c value       Certificate signing request filename, if an external CSR is to be used
    --server value, -s value    CA hostname (and optionally :port). The server certificate must be trusted in order to avoid further modifications to the client. (default: "https://acme-v02.api.letsencrypt.org/directory")
-   --email value, -m value     Email used for registration and recovery contact.
-   --filename value            Filename of the generated certificate
    --accept-tos, -a            By setting this flag to true you indicate that you accept the current Let's Encrypt terms of service.
+   --email value, -m value     Email used for registration and recovery contact.
+   --csr value, -c value       Certificate signing request filename, if an external CSR is to be used
    --eab                       Use External Account Binding for account registration. Requires --kid and --hmac.
    --kid value                 Key identifier from External CA. Used for External Account Binding.
    --hmac value                MAC key from External CA. Should be in Base64 URL Encoding without padding format. Used for External Account Binding.
    --key-type value, -k value  Key type to use for private keys. Supported: rsa2048, rsa4096, rsa8192, ec256, ec384 (default: "rsa2048")
+   --filename value            Filename of the generated certificate
    --path value                Directory to use for storing the data (default: "./.lego")
    --exclude value, -x value   Explicitly disallow solvers by name from being used. Solvers: "http-01", "dns-01", "tls-alpn-01".
+   --http-timeout value        Set the HTTP timeout value to a specific value in seconds. The default is 10 seconds. (default: 0)
    --webroot value             Set the webroot folder to use for HTTP based challenges to write directly in a file in .well-known/acme-challenge
    --memcached-host value      Set the memcached host(s) to use for HTTP based challenges. Challenges will be written to all specified hosts.
    --http value                Set the port and interface to use for HTTP based challenges to listen on. Supported: interface:port or :port
    --tls value                 Set the port and interface to use for TLS based challenges to listen on. Supported: interface:port or :port
    --dns value                 Solve a DNS challenge using the specified provider. Disables all other challenges. Run 'lego dnshelp' for help on usage.
-   --http-timeout value        Set the HTTP timeout value to a specific value in seconds. The default is 10 seconds. (default: 0)
-   --dns-timeout value         Set the DNS timeout value to a specific value in seconds. The default is 10 seconds. (default: 0)
-   --disable-cp                By setting this flag to true, disables the need to wait the propagation of the TXT record to all authoritative name servers.
+   --dns-disable-cp            By setting this flag to true, disables the need to wait the propagation of the TXT record to all authoritative name servers.
    --dns-resolvers value       Set the resolvers to use for performing recursive DNS queries. Supported: host:port. The default is to use the system resolvers, or Google's DNS resolvers if the system's cannot be determined.
+   --dns-timeout value         Set the DNS timeout value to a specific value in seconds. Used only when performing authoritative name servers queries. The default is 10 seconds. (default: 0)
    --pem                       Generate a .pem file by concatenating the .key and .crt files together.
    --help, -h                  show help
    --version, -v               print the version
+
 ```
 
 ### Sudo
