+++ cmd/flags.go
@@ -1,6 +1,9 @@
 package cmd
 
-import "github.com/urfave/cli"
+import (
+	"github.com/urfave/cli"
+	"github.com/xenolf/lego/acme"
+)
 
 func CreateFlags(defaultPath string) []cli.Flag {
 	return []cli.Flag{
@@ -9,25 +12,21 @@ func CreateFlags(defaultPath string) []cli.Flag {
 			Usage: "Add a domain to the process. Can be specified multiple times.",
 		},
 		cli.StringFlag{
-			Name:  "csr, c",
-			Usage: "Certificate signing request filename, if an external CSR is to be used",
-		},
-		cli.StringFlag{
 			Name:  "server, s",
-			Value: "https://acme-v02.api.letsencrypt.org/directory",
+			Value: acme.LEDirectoryProduction,
 			Usage: "CA hostname (and optionally :port). The server certificate must be trusted in order to avoid further modifications to the client.",
+		},
+		cli.BoolFlag{
+			Name:  "accept-tos, a",
+			Usage: "By setting this flag to true you indicate that you accept the current Let's Encrypt terms of service.",
 		},
 		cli.StringFlag{
 			Name:  "email, m",
 			Usage: "Email used for registration and recovery contact.",
 		},
 		cli.StringFlag{
-			Name:  "filename",
-			Usage: "Filename of the generated certificate",
-		},
-		cli.BoolFlag{
-			Name:  "accept-tos, a",
-			Usage: "By setting this flag to true you indicate that you accept the current Let's Encrypt terms of service.",
+			Name:  "csr, c",
+			Usage: "Certificate signing request filename, if an external CSR is to be used",
 		},
 		cli.BoolFlag{
 			Name:  "eab",
@@ -47,6 +46,10 @@ func CreateFlags(defaultPath string) []cli.Flag {
 			Usage: "Key type to use for private keys. Supported: rsa2048, rsa4096, rsa8192, ec256, ec384",
 		},
 		cli.StringFlag{
+			Name:  "filename",
+			Usage: "Filename of the generated certificate",
+		},
+		cli.StringFlag{
 			Name:  "path",
 			Usage: "Directory to use for storing the data",
 			Value: defaultPath,
@@ -54,6 +57,10 @@ func CreateFlags(defaultPath string) []cli.Flag {
 		cli.StringSliceFlag{
 			Name:  "exclude, x",
 			Usage: "Explicitly disallow solvers by name from being used. Solvers: \"http-01\", \"dns-01\", \"tls-alpn-01\".",
+		},
+		cli.IntFlag{
+			Name:  "http-timeout",
+			Usage: "Set the HTTP timeout value to a specific value in seconds. The default is 10 seconds.",
 		},
 		cli.StringFlag{
 			Name:  "webroot",
@@ -75,21 +82,17 @@ func CreateFlags(defaultPath string) []cli.Flag {
 			Name:  "dns",
 			Usage: "Solve a DNS challenge using the specified provider. Disables all other challenges. Run 'lego dnshelp' for help on usage.",
 		},
-		cli.IntFlag{
-			Name:  "http-timeout",
-			Usage: "Set the HTTP timeout value to a specific value in seconds. The default is 10 seconds.",
-		},
-		cli.IntFlag{
-			Name:  "dns-timeout",
-			Usage: "Set the DNS timeout value to a specific value in seconds. The default is 10 seconds.",
-		},
 		cli.BoolFlag{
-			Name:  "disable-cp",
+			Name:  "dns-disable-cp",
 			Usage: "By setting this flag to true, disables the need to wait the propagation of the TXT record to all authoritative name servers.",
 		},
 		cli.StringSliceFlag{
 			Name:  "dns-resolvers",
 			Usage: "Set the resolvers to use for performing recursive DNS queries. Supported: host:port. The default is to use the system resolvers, or Google's DNS resolvers if the system's cannot be determined.",
+		},
+		cli.IntFlag{
+			Name:  "dns-timeout",
+			Usage: "Set the DNS timeout value to a specific value in seconds. Used only when performing authoritative name servers queries. The default is 10 seconds.",
 		},
 		cli.BoolFlag{
 			Name:  "pem",
